Harden order edits after PR review
Constraint: CodeRabbit flagged precision, rollback, and stale-fill review risks after PR #102 opened.
Rejected: Adding validation dependencies for a single request DTO | Existing service/CoreException validation covers null, non-positive, and overflowed prices without widening dependencies.
Rejected: Continuing guarded claim misses | It can fill lower-priority stale candidates before a fresh price event re-sorts edited orders.
Confidence: high
Scope-risk: moderate
Directive: Keep API boundary prices precise while preserving the existing double-based trading domain until a broader money-type migration is planned.
Tested: cd backend && ./gradlew test --tests '*ModifyOrderServiceTest' --tests '*MarketOrderExecutionServiceTest' --console=plain; cd backend && ./gradlew check --console=plain; npm test --workspace frontend; npm run build --workspace frontend; npm run lint; git diff --check

+++ backend/src/main/java/coin/coinzzickmock/feature/order/application/result/ModifyOrderResult.java
@@ -1,12 +1,14 @@
 package coin.coinzzickmock.feature.order.application.result;
 
+import java.math.BigDecimal;
+
 public record ModifyOrderResult(
         String orderId,
         String symbol,
         String status,
-        Double limitPrice,
+        BigDecimal limitPrice,
         String feeType,
-        double estimatedFee,
-        double executionPrice
+        BigDecimal estimatedFee,
+        BigDecimal executionPrice
 ) {
 }
